+++ src/Date.ts
@@ -2,7 +2,8 @@
  * @since 2.0.0
  */
 import { IO } from './IO'
-import { Eq } from './Eq'
+import { Eq, fromEquals } from './Eq'
+import { Ord, fromCompare, ordNumber } from './Ord'
 
 /**
  * Returns the current `Date`
@@ -19,22 +20,36 @@ export const create: IO<Date> = () => new Date()
 export const now: IO<number> = () => new Date().getTime()
 
 /**
- * @since 2.6.0
+ * @since 3.0.0
  */
-export const eqDate: Eq<Date> = {
-  equals: (x, y) => x.getDate() === y.getDate()
-}
+export const eqDate: Eq<Date> =
+  /*#__PURE__*/
+  fromEquals((x, y) => x.valueOf() === y.valueOf())
 
 /**
- * @since 2.6.0
+ * @since 3.0.0
  */
-export const eqMonth: Eq<Date> = {
-  equals: (x, y) => x.getMonth() === y.getMonth()
-}
+export const eqGetDate: Eq<Date> =
+  /*#__PURE__*/
+  fromEquals((x, y) => x.getDate() === y.getDate())
 
 /**
- * @since 2.6.0
+ * @since 3.0.0
  */
-export const eqYear: Eq<Date> = {
-  equals: (x, y) => x.getFullYear() === y.getFullYear()
-}
+export const eqGetMonth: Eq<Date> =
+  /*#__PURE__*/
+  fromEquals((x, y) => x.getMonth() === y.getMonth())
+
+/**
+ * @since 3.0.0
+ */
+export const eqGetFullYear: Eq<Date> =
+  /*#__PURE__*/
+  fromEquals((x, y) => x.getFullYear() === y.getFullYear())
+
+/**
+ * @since 3.0.0
+ */
+export const ordDate: Ord<Date> =
+  /*#__PURE__*/
+  fromCompare((x, y) => ordNumber.compare(x.valueOf(), y.valueOf()))