+++ test/Date.ts
@@ -16,17 +16,29 @@ describe('Date', () => {
   })
 
   it('eqDate', () => {
-    assert.deepStrictEqual(_.eqDate.equals(new Date(2000, 10, 1), new Date(2000, 11, 1)), true)
-    assert.deepStrictEqual(_.eqDate.equals(new Date(2000, 10, 1), new Date(2000, 11, 2)), false)
+    assert.deepStrictEqual(_.eqDate.equals(new Date(0), new Date(0)), true)
+    assert.deepStrictEqual(_.eqDate.equals(new Date(0), new Date(1)), false)
+    assert.deepStrictEqual(_.eqDate.equals(new Date(1), new Date(0)), false)
   })
 
-  it('eqMonth', () => {
-    assert.deepStrictEqual(_.eqMonth.equals(new Date(2000, 10, 1), new Date(2000, 10, 7)), true)
-    assert.deepStrictEqual(_.eqMonth.equals(new Date(2000, 10, 1), new Date(2000, 11, 7)), false)
+  it('eqGetDate', () => {
+    assert.deepStrictEqual(_.eqGetDate.equals(new Date(2000, 10, 1), new Date(2000, 11, 1)), true)
+    assert.deepStrictEqual(_.eqGetDate.equals(new Date(2000, 10, 1), new Date(2000, 11, 2)), false)
   })
 
-  it('eqYear', () => {
-    assert.deepStrictEqual(_.eqYear.equals(new Date(2000, 10, 1), new Date(2000, 11, 7)), true)
-    assert.deepStrictEqual(_.eqYear.equals(new Date(2000, 10, 1), new Date(2001, 11, 7)), false)
+  it('eqGetMonth', () => {
+    assert.deepStrictEqual(_.eqGetMonth.equals(new Date(2000, 10, 1), new Date(2000, 10, 7)), true)
+    assert.deepStrictEqual(_.eqGetMonth.equals(new Date(2000, 10, 1), new Date(2000, 11, 7)), false)
+  })
+
+  it('eqGetFullYear', () => {
+    assert.deepStrictEqual(_.eqGetFullYear.equals(new Date(2000, 10, 1), new Date(2000, 11, 7)), true)
+    assert.deepStrictEqual(_.eqGetFullYear.equals(new Date(2000, 10, 1), new Date(2001, 11, 7)), false)
+  })
+
+  it('ordDate', () => {
+    assert.deepStrictEqual(_.ordDate.compare(new Date(0), new Date(0)), 0)
+    assert.deepStrictEqual(_.ordDate.compare(new Date(0), new Date(1)), -1)
+    assert.deepStrictEqual(_.ordDate.compare(new Date(1), new Date(0)), 1)
   })
 })